+++ R/methods.R
@@ -16,6 +16,11 @@ renv_methods_map <- function() {
     renv_file_broken = c(
       unix  = "renv_file_broken_unix",
       win32 = "renv_file_broken_win32"
+    ),
+
+    renv_paths_sandbox = c(
+      unix  = "renv_paths_sandbox_unix",
+      win32 = "renv_paths_sandbox_win32"
     )
 
   )